add alertdialog when logout, withdraw add progress when login

+++ app/src/main/java/com/mngs/kimyounghoon/mngs/home/HomeFragment.kt
@@ -4,12 +4,15 @@ import android.databinding.DataBindingUtil
 import android.os.Bundle
 import android.os.Handler
 import android.support.design.widget.TabLayout
+import android.support.v7.app.AlertDialog
 import android.view.*
 import com.google.firebase.auth.FirebaseAuth
 import com.mngs.kimyounghoon.mngs.AbstractFragment
 import com.mngs.kimyounghoon.mngs.AccountManager
 import com.mngs.kimyounghoon.mngs.R
+import com.mngs.kimyounghoon.mngs.SingleLiveEvent
 import com.mngs.kimyounghoon.mngs.databinding.FragmentHomeBinding
+import com.mngs.kimyounghoon.mngs.utils.setupProgressDialog
 
 
 class HomeFragment : AbstractFragment(), TabLayout.OnTabSelectedListener {
@@ -26,7 +29,7 @@ class HomeFragment : AbstractFragment(), TabLayout.OnTabSelectedListener {
 
     private lateinit var fragmentHomeBinding: FragmentHomeBinding
     private lateinit var homeFragmentStatePagerAdapter: HomeFragmentStatePagerAdapter
-
+    private var needProgress = SingleLiveEvent<Boolean>()
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setHasOptionsMenu(true)
@@ -48,6 +51,11 @@ class HomeFragment : AbstractFragment(), TabLayout.OnTabSelectedListener {
             locateListener?.openWriteLetter()
         }
         return fragmentHomeBinding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        view.setupProgressDialog(this,needProgress)
     }
 
     private fun setUpTabLayout() {
@@ -86,20 +94,31 @@ class HomeFragment : AbstractFragment(), TabLayout.OnTabSelectedListener {
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean = when (item.itemId) {
         R.id.menu_logout -> {
-            FirebaseAuth.getInstance().signOut()
-            locateListener?.openLogin()
+            context?.apply {
+                AlertDialog.Builder(this).setTitle(R.string.logout_alert_text).setPositiveButton(R.string.yes) { _, _ ->
+                    FirebaseAuth.getInstance().signOut()
+                    locateListener?.openLogin()
+                }.setNegativeButton(R.string.no, null).show()
+            }
             true
         }
         R.id.menu_withdraw -> {
-            val user = FirebaseAuth.getInstance().currentUser
-            user?.delete()
-                    ?.addOnCompleteListener {
-                        if (it.isSuccessful) {
-                            locateListener?.openLogin()
-                        } else {
-                            //todo 탈퇴 실패
-                        }
-                    }
+            context?.apply {
+                AlertDialog.Builder(this).setTitle(R.string.withdraw_alert_text).setPositiveButton(R.string.yes) { _, _ ->
+                    needProgress.value = true
+                    val user = FirebaseAuth.getInstance().currentUser
+                    user?.delete()
+                            ?.addOnCompleteListener {
+                                if (it.isSuccessful) {
+                                    needProgress.value = false
+                                    locateListener?.openLogin()
+                                } else {
+                                    needProgress.value = false
+                                    //todo 탈퇴 실패
+                                }
+                            }
+                }.setNegativeButton(R.string.no, null).show()
+            }
             true
         }
         else -> {
